[#83447690] Bring toolbox under the new delete
area system.  Supposed to be an implementation of this upstream change:
https://github.com/google/blockly/commit/c9169285d94cc68c99b91b9bd8d3606
96d0f3f09
In reality, we’re diverging a bit and doing things better.
This commit leaves in a bug that allows you to delete blocks that you
shouldn’t be able to delete.  That will be fixed on this branch before
anything gets merged.