Maze: don't serve animated gifs as pngs.
PvZ maze had two .png files that were actually animated .gif files, and IE11 refused to show them.  Angry Birds maze had one .png file in a similar state.  This change allows these levels to specify a .gif file for these cases.

It was a bit confusing that we used obstacle.png for the idle animation and obstacle.gif for the active animation, so I renamed obstacle.png to obstacleIdle.gif, etc., for this special-casing.