Fix Redis API object construction.
Config option CDO.geocoder_redis_url must be passed to Redis.new
as the :url parameter, not the :host parameter, because it
also includes the port number.  This seems to be a less popular
way to construct the Redis API, but it's how our current redis
configuration works.

See https://github.com/redis/redis-rb#getting-started

[ci skip]