Fix the NetSim activity monitor test.
The test was invalidated when Sarah F. changed netsim.script and
added some public-facing "offline" levels.  The test was checking
the first level in the script, which was no longer an actual
NetSim level.

Now, the test uses the "allthethings" script that we've set up just
for test purposes, and searches for the first actual NetSim level.
The test passes when running against a NetSim level.