Fix "Changing Parameter Names" UI test.
Test started failing because the level changed out from under it.
Copied a version of the level from a few days ago (a passing one)
into the allthethings script so it won't change in the future.

Updated the test to pass (doesn't need the cancel button presses anymore).
Aliased involved blocks so when it _does_ break again, it will
be easier to read what it was trying to do.