Correct shard disconnect.
Adds a proper disconnect/unsubscribe step for the shard, which
is triggered when navigating away or when the shard is reset.

Before we kept the shard around unchanged, and just replaced it on
reconnect (an artifact of an old attempt at an auto-reconnect system).

This didn't work anymore because our PubSub subscription steps are
tied to shard creation, and we need to unsubscribe properly so we
don't build up multiple subscriptions across shard resets.

Now we properly unsubscribe/resubscribe PubSub (and use some
previously unused code to this effect, go figure).