Remove wire update after connection to router.
Before the client connected to the router, and then the router
updated the wire with addresses and hostnames.  Now the client
creates the wire with all information in place (including an
address that we _think_ is available) so if connection is successful,
no further updates are necessary.

Required quite a few changes to tests - I think I preserved the
spirit of the tests that got changed/removed.