HUGE refactor; splits frequency chart into two
We now have three interactable rows and two full charts. The top chart
and top two rows are essentially what we had before, the bottom chart
and third now are a new "unassigned letters" category.

This means that we no longer have to rely on just switching to assign
substitutions; we actually start with nothing assigned, and drag
assignments up from row 3 to row 2