NetSimApi provides more detailed error information.
The server api always generates a status 400 on failed requests,
but can optionally provide more detailed error information.

Initial use case for this is for reporting routerNumber conflicts
on insert, and requests rejected because the strict router limit
has been reached.

The client gets a special NetSimApiError type that can parse and
provide the additional error information, and the router creation
flow in particular now has a special user-visible warning  when
the router limit has been reached, and knows not to display a
user-facing error when a conflict occurs (because they probably
saw a router added anyway - it just happened to be added by
another user at about the same time).