Use real Dashboard schema in Pegasus tests.
Uses ActiveRecord and sqlite3 to load the real Dashboard schema
for use during Pegasus tests - so the only custom code we write
is for putting test data into that database (which we might be
able to use Factory Girl for in the future).

Switched from using a sqlite in-memory database to using a filesystem
database, because I couldn't get multiple connections to an in-memory
database.  It seems sqlite3 /should/ support this, but ActiveRecord
and Sequel do not.  Instead we create a database file, and we remove
it after the tests run (if it was created at all) with an extension
to the rake test task.