Positioning start blocks using Blockly
Previously, the initial position of blockly blocks was calculated in
studio code. Because this was happening before the blocks were rendered,
the code had to be extremely clever and yet was still causing
undesirable positioning in several cases (most notably when some of the
start blocks were significantly tall).

This PR moves the logic for positioning these blocks into blockly, where
it can use the rendered height of each block to calculate positioning,
and can therefore be much more straightforward.