hoc2015: draw walls correct using either level.wallMap or level.map.
Now draws walls correctly for any of the following situations:
  - level.map specifies random tiles, drawn once at level load. (e.g. level 1)
  - level.wallMap specifies a skin layout, drawn upfront and redrawn when background changed, if necessary. (e.g. level 14)
  - a hybrid in which level.map specifies sprite/item locations, while level.wallMap also specifies a skin-based wall layout. (e.g. level 7)
  - when level.wallMap is set and a student calls setMap.

It gets rid of the defaut skin wall map in hoc2015 skin ("blank") because this made the first situation difficult to deal with.  Now, when there is no skin-based map specified, setMap doesn't do anything, so it doesn't go about erasing the random tiles that are drawn once-only.  (They are drawn once-only to avoid having the randomly-chosen tiles change.  An alternative would have been to use a non-truly-random layout.)

There is no longer