Double timing for more reliable coalescing test
The tests in question depend on execution time (and do
factor that into their design) but can fail on particularly
slow test runners.  This change doubles the coalescing window
from 50ms to 100ms and doubles the corresponding test delays,
to make the test more stable in slow environments
(e.g. CircleCI) at the cost of taking slightly longer to run.