Fixes Blockly Positioning of "arranged" blocks
Blocks that had "arrangments" that only specified X or Y coordinates
(not both) were writing both X and Y coordinates to the XML, causing
occasional `undefined` coordinates.

Additionally, blocks with X coordinates set in this way were being
positioned absolutely with Y=0. This was highly undersirable, as
Category blocks have an arrangement that positions them all this way.
The end result of all this mess was that category blocks were being
stacked up at the top of the toolbox editor, making them all but
useless.