Improve FunOMeter test stability
The Fun-O-Meter test has been chronically flaky on iOS devices since its
creation. The flakiness had - as far as I can tell - two  root causes:

- As part of the test, we complete the level we're on and advance to the
  next level before revisiting the original level. In the original test,
  the "next" level had an autoplaying YouTube video, which the iPad
  simulator on SauceLabs had some problems with. I rewrote the test to
  use a different level that's just as easy to solve and progresses to a
  level WITHOUT a video.

- The test was advancing to the next page so that a required AJAX
  request could complete, but was not actually waiting to make sure it
  had done so before returning. It now does.

There also may or may not have been a problem with user signin being a
bit flaky itself. This appears to no longer be the case; Brad's
change to the student signin process is likely to thank for that.