First draft of randomized conditionals
Adds a set of new syntax to Bee levels for randomized conditional cells;
that is, cells that can contain one of a number of kinds of content.
This will let the cells hidden by clouds be a true mystery.

This also means that levels that use these cells will have a number of
possible grids; if that's the case, we run the user's code against all
possible grids. If their code succeeds on all, then we randomly select
one of the grids to animate on top of. If their code fails on any, we
randomly select from one of the failing grids to display.