Fix header issues (more generally) when strings are too long
In some translations, the nav strings are too long, causing the nav links
to be pushed beneath the language dropdown. In a previous check-in, we
simply adjusted the margins and hoped it would work for most languages.

This change addresses the issue more generically. It caps the width of
each individual link. When they're too wide, text is clipped and
ellipses are added. Title attributes are also added for each link,
so users can see the full text in a tooltip for any of the clipped
strings.