Repair Studio setSpriteSize
A recent refactor changed the way setSpriteSize works when used by the
user at runtime; namely, it no longer does. Specifically, it no longer
does because setSpriteSize was changing the size property on the Sprite
object and then calling setSprite to reset the image. However, setSprite
calls setImage and setLegacy image, which check to see if the image has
actually changed and return without doing anything if it has not.

Because the image was unchanged but the size was, this means Sprite
ended up in a state where its internal size property was changing, as
were a few things that relied on size (like display position), but not
the displayed size of the image itself.

A more surgical solution would be to refactor setSize as a method on
Sprite itself, that then handles updating all the properties in a more
precise and ordered way, rather than simply blowing away everything and
rebuilding it on each change. However, that's a much larger task for a
later time. For now, we just force setSpriteSize to blow away everything
so that it can be actually rebuild and not just partially rebuilt.