Adds a React Grid Editor for Karel levels
For Karel levels (as in, not for plain Grid levels like basic mazes),
replaces the existing hansontable excel-style grid editor with a custom
React component that uses a smart Cell editor.

This allows us to move away from the world of representing grids via
arrays of syntactically-significant strings, and toward a more complex
(but also more straightforward) system of JSON serialization.