Muti submit/unsubmit works.  submit is field on UserLevel.
While the dialog is hardcoded English, multi submit/unsubmit flow now works.  I've followed the suggestion to stop overloading result==1000 as submitted, and instead UserLevel has a submitted boolean property.  Since only the server knows when a level is submitted or not, the client-side clientState cache doesn't need to know about it.  The dot in the level progress in the header, and in the dropdown, turns purple when a level is in the submitted state.