Adds concept of "unused" blocks to Blockly
"unused" blocks are blocks that are not intended to be top blocks (aka
blocks that have a "previous" connection) but have no parents in the
main block space.

Previously, these blocks were considered to be errors by Studio. Now,
they have a special render pathway that highlights for the user the fact
that they are unattached and therefore unused, and Studio deals with
them appropriately.