Add test that validates the data pulled from cdo-donors
cdo-donors has a weight column whose maximum value must be 1.0 in order
for our weighted random select to work. Even though the gsheet that this
data is coming from has algorithms that should enforce this, it's gotten
messed up on occasion, and that mistake has in at least one case made it
through to production.

No longer!