Swallow p5 library exception on remove()
Bug has been fixed in p5.js and we just need to update our version
of the library.
Submitting this hack to swallow the error in case I don't get to it
today (we have a hackathon tomorrow and this would be a pain to live
with.
See https://github.com/processing/p5.js/pull/1054 for bugfix.