Change generated level url to absolute rather than relative
The server-generated level summaries include relative level urls, but
the summaries are used in some instances by pegasus (most notably the
teacher dashboard), which obviously requires an absolute url.

This PR also updates the application config to set the default url port
when in development mode, so generated urls will work in a development
environment.