Refactor authored hints workflow
Originally, clicking the small instructions area would display the
Instructions modal dialog.

Then, we added authored hints to the system, and changed it so clicking
the small instructions area would display a qtip dialog offering you the
choice to either view the instructions or view a hint. If you viewed the
instructions, you'd also see a list of your already-seen hints and a
button that would add one more to the list. If you chose to view the
hint, the content of the qtip dialog would be replaced with a hint.

Now, we've removed the intermediary dialog that presents you with a
choice. Instead, clicking the small instructions area displays the
instructions dialog, and clicking the "see hint" button in the
instructions dialog displays the hint in a qtip dialog.