Link to levelbuilder from admin panel
The "edit" link that appears in the admin panel on production is
useless, since we don't allow for levels to be edited on production.

Now, the admin panel will check to see if the server supports level
editing. If it doesn't and if you are viewing a script level (rather
than just a level at `/levels/{id}`, it will link you to the
levelbuilder versions of that script level, from which you can get an
"edit" button that actually works.