Remove deprecated Readonly Blockspace styles
Originally introduced in 930c330042b35058da5f1520711e90f901a0d81e, these
styles on readonly codeWorkspaces are no longer relevant.

We currently have three kinds of readonly workspaces:

1. Teacher views of student work and solutions
2. Feedback blocks (both in the feedback dialog and in authored hints)
3. Embedded code in multi levels

The first we actually DON'T want to see hidden blocks, and the border is
overwritten by javascript styling anyway.

The second doesn't use the `#codeWorkspace` ID since it's not in the
code workspace, so these styles don't apply anyway.

The third doesn't add the `.readonly` class to the codeworkspace but
rather to the `body` tag of the iframe. This should probably be updated
to use an embedded blockspace like the second so we have fewer cases.