district dropdown: now features all districts per state.
This adds a .tsv file generated from our 18,000-long gsheet of US districts.  It's imported into the database.

The district dropdown then retrieves the appropriate set of per-state entries when a state is selected in the preceding dropdown, from which the user can select their district.

Definitely work in progress, but wanted to get this saved.