Log to new relic when not using watch
To get this to work, I had to change a bunch of things around in
time-grunt-nowatch. I also had to switch to synchronous file reading so
that things would happen in the right order. I also had to pass in a log
function in order to get anything to show up on the console :(

At least I think it actually really really works now.