district dropdown: lots more state transitions.
The small district dropdown widget actually involves a bunch of state transitions, as the various options are explored.  This change hopefully takes care of all of them, getting the right information fed into the form, and clearing out any entries that shouldn't make it through based on what the user ultimately chooses.