Include babel-polyfill at entry points [test ui]
We've been including babel-polyfill in files that need it
(inconsistently) to get access to Promises and Object.assign,
but there are still pockets of code that don't have access
to it.  This moves the include to our package entry points,
as recommended here:

https://babeljs.io/docs/usage/polyfill/#usage-in-node-browserify-webpack

...which makes it available everywhere.

I've included at the top of appMain.js instead of the many main.js
files, because appMain.js is the first include in every main.js file.

I'm including at the top of code-studio.js even though we don't use
it there yet to provide full support in most code-studio code.

I've omitted the polyfill from some of our specialized bundles
(like react-only.js) to keep their size down.