Initial working version of drag out to delete an element from app space
This includes removing some CSS styles that were checked in and
pushed up earlier for Dave to do an initial debug. These styles are
no longer needed to make the feature work, so was cleaned up.

Logic for detecting whether a position is in bounds was also refactored
out and placed in gridUtils for multiple components to share and use.