Generate presigned URL for non-public uploads
In the case of build logs we don't want to make them public, but we do want to provide a link in Slack.  The solution to this is to generate a "presigned" url which is only valid for a limited time.

Also fixes a bug in the truncated log for Slack.