Better phantomjs dependency in blockly-core
Instead of depending on Will's custom build of phantomjs>2 (which we've moved away from in other parts of our code), depend on the phantomjs-prebuilt package which has been fairly reliable lately.

Also, invoke phantomjs via `npm bin`/phantomjs instead of the full installed path inside node_modules, which we really shouldn't have any knowledge of (and which can change - as in this particular package change).