signup: navigate to right URLs.  code review feedback.  [test ui]
As well as incorporating code review feedback, this supports properly going to the proper URL after a correct sign up (and therefore sign in)... a return_to URL stored in the session if possible, otherwise a teacher or student URL as appropriate.  This change deletes the return_to URL from the session when the page is first visited.  It's a small change in existing behaviour, but it will actually reduce confusion incidents in which somebody signs up weeks later and is redirected to a page that they had tried to visit long ago.